# customers/internal/rest/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Customers",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "CustomersService"
    }
  ],
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/customers": {
      "post": {
        "summary": "Create a new customer",
        "operationId": "createCustomer",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/customerspbRegisterCustomerResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/customerspbRegisterCustomerRequest"
            }
          }
        ],
        "tags": [
          "Customer"
        ]
      }
    },
    "/api/customers/{id}": {
      "get": {
        "summary": "Get a customer",
        "operationId": "getCustomer",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/customerspbGetCustomerResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Customer"
        ]
      }
    },
    "/api/customers/{id}/disable": {
      "put": {
        "summary": "Disable a customer",
        "operationId": "DisableCustomer",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/customerspbDisableCustomerResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "tags": [
          "Customer"
        ]
      }
    },
    "/api/customers/{id}/enable": {
      "put": {
        "summary": "Enable a customer",
        "operationId": "enableCustomer",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/customerspbEnableCustomerResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "tags": [
          "Customer"
        ]
      }
    }
  },
  "definitions": {
    "customerspbAuthorizeCustomerResponse": {
      "type": "object"
    },
    "customerspbCustomer": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "smsNumber": {
          "type": "string"
        },
        "enabled": {
          "type": "boolean"
        }
      }
    },
    "customerspbDisableCustomerResponse": {
      "type": "object"
    },
    "customerspbEnableCustomerResponse": {
      "type": "object"
    },
    "customerspbGetCustomerResponse": {
      "type": "object",
      "properties": {
        "customer": {
          "$ref": "#/definitions/customerspbCustomer"
        }
      }
    },
    "customerspbRegisterCustomerRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "smsNumber": {
          "type": "string"
        }
      }
    },
    "customerspbRegisterCustomerResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}